Task 1: Implement AQueue front() Accessor Method is compiling and running tests.
On this task 1 we added access front of queue
Accessor method to get a copy of the item currently at the front of this queue. T Returns a copy of the front item of the queue.Also, uncommented test a queue function.we also added member methods for managing the Queue internally.